+++ Libplanet/Action/ActionTypeAttribute.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Reflection;
+using Libplanet.Tx;
 
 namespace Libplanet.Action
 {
@@ -40,11 +41,20 @@ namespace Libplanet.Action
         /// <c>null</c>.</returns>
         public static string ValueOf(Type actionType)
         {
-            return actionType
+            string typeIdentifier = actionType
                 .GetCustomAttributes()
                 .OfType<ActionTypeAttribute>()
                 .Select(attr => attr.TypeIdentifier)
                 .FirstOrDefault();
+
+            if (typeIdentifier is null)
+            {
+                throw new InvalidActionTypeException(
+                    "Action type should be annotated with ActionTypeAttribute."
+                );
+            }
+
+            return typeIdentifier;
         }
     }
 }